Support react-native modals (#7156)
Currently, using react-native declarative modals can cause view controllers to be detached from the view hierarchy. For example, when presenting an RNN modal from react-native modal and then dismissing the react-native modal, the RNN modal on top of it remain detached as its holder is dismissed. It can also cause the new RNN modal to be dismissed.

In this PR we implemented `RCTModalHostViewManager.presentationBlock`  and `RCTModalHostViewManager.dismissalBlock` in order to have the same modals logic we have in RNN.