Fix remx optional dependency (#7241)
Currently, remx is an optional peer dependency that is required for our mock consumers only. But it is actually being imported from our main `index.ts` which break users who don't import it even though they are not using mocks.
The solution is to change the way users import our mocks and make it explicit. 

Mocks now should be imported from `react-native-navigation/Mock`.


Co-authored-by: Carl-Gerhard Lindesvärd <c.lindesvard@klarna.com>